requirements.txt
albumentations==1.3.0
asynctest==0.13.0
cityscapesscripts==2.2.1
ffmpegcv==0.2.6
imagecorruptions==1.1.2
imageio==2.25.0
instaboostfast==0.1.2
kwarray==0.6.8
matplotlib==3.6.3
memory_profiler==0.61.0
mmcv==1.7.1
mmcv_full==1.7.1
mmtrack==0.14.0
model_archiver==1.0.3
numpy==1.23.5
onnx==1.13.0
onnxruntime==1.13.1
onnxsim==0.4.13
opencv_python==4.7.0.68
Pillow==9.4.0
psutil==5.9.4
pycocotools==2.0.6
pytest==7.2.1
pytorch_sphinx_theme==0.0.19
PyYAML==6.0
requests==2.28.1
scikit_learn==1.2.1
scipy==1.10.0
seaborn==0.12.2
setuptools==65.6.3
six==1.16.0
streamlit==1.17.0
streamlit_webrtc==0.44.2
tensorrt==8.5.2.2
terminaltables==3.1.10
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
ts==0.5.1
wandb==0.13.9
xlrd==2.0.1
xlutils==2.0.0
